refactor: Move much of ui-keyring/options into Keyring class. Fix Keyring instance assertion
* move initOptions from ui-keyring/options/index into index.ts

* move createOptionHeader from ui-keyring/options/header keyring.createOptionHeader

* move optionsSubject, addAccounts, addAddresses, emptyOptions from ui-keyring/options/index into Keyring class so call with keyring.optionsSubject

* alphabetical ordering of isAvailable and isPassValid

* did not move createItem from ui-keyring/src/options/item.tsx (which is imported as makeOption) into index.ts, since it uses React

* fix KeyringInstance singleton assertion
