+++ packages/ui-app/src/InputAddress/index.tsx
@@ -10,10 +10,8 @@ import './InputAddress.css';
 import React from 'react';
 import store from 'store';
 import keyring from '@polkadot/ui-keyring/index';
-import createOptionHeader from '@polkadot/ui-keyring/options/header';
 import addressDecode from '@polkadot/util-keyring/address/decode';
 import addressEncode from '@polkadot/util-keyring/address/encode';
-import { optionsSubject } from '@polkadot/ui-keyring/options';
 import makeOption from '@polkadot/ui-keyring/options/item';
 import withObservableBase from '@polkadot/ui-react-rx/with/observableBase';
 import isHex from '@polkadot/util/is/hex';
@@ -191,7 +189,7 @@ class InputAddress extends React.PureComponent<Props, State> {
       if (publicKey.length === 32) {
         if (!matches.find((item) => item.key === RECENT_KEY)) {
           matches.push(
-            createOptionHeader('Recent')
+            keyring.createOptionHeader('Recent')
           );
         }
 
@@ -216,4 +214,4 @@ class InputAddress extends React.PureComponent<Props, State> {
 export { InputAddress };
 
 // @ts-ignore There are still some issues with props and types - this is valid
-export default withObservableBase(optionsSubject, { propName: 'optionsAll' })(InputAddress);
+export default withObservableBase(keyring.optionsSubject, { propName: 'optionsAll' })(InputAddress);
